Book: Fix image, move text related to building book into external README (#30)
* Book: Fix image, move text related to building book into external README

It looks like we can't link to local items in the book that are outside
of the `src` directory. I've moved the `img` directory under `src` to
amend this.

The intro page previously contained a "Documentation" section which
explains how to build the mdbook. These steps aren't relevant to readers
of `fuel-indexer` who are just trying to familiarise with `fuel-indexer`
itself. I've moved this text into the book's root directory, just like
we have for the Sway book's directory [here][1].

Also removes the line that recommends reading the book, as the reader is
already reading the book.

[1]: https://github.com/FuelLabs/sway/blob/master/docs/README.md

* Fix path to img in top-level README