Consolidate pools across application
While trying to fix the SQLite breakage, I noticed that the
max_connections() setting was not being respected. After some
investigation, I realized that we create multiple pools across
the application instead of using one pool and acquiring the
connections from there. This commit fixes that and brings
the application more in line with the use case of a database pool.